+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `Remove-ZeroedEnums`
+- `Test-IsZeroedEnum`
   - Added private function to remove enums with a zero value from a hashtable.
 
 ### Changed