+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `Clear-ZeroedEnumPropertyValue`
+  - Fixed enum type check.
+
 ## [1.2.0] - 2025-02-02
 
 ### Added