+++ CHANGELOG.md
@@ -5,6 +5,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Wiki
+  - How to use this base class. Fixes [#5](https://github.com/dsccommunity/DscResource.Base/issues/5)
+  and [#19](https://github.com/dsccommunity/DscResource.Base/issues/19).
+
 ## [1.3.0] - 2025-03-15
 
 ### Added