+++ CHANGELOG.md
@@ -10,6 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Wiki
   - How to use this base class. Fixes [#5](https://github.com/dsccommunity/DscResource.Base/issues/5)
   and [#19](https://github.com/dsccommunity/DscResource.Base/issues/19).
+  - Add WikiContent to release assets.
+
+### Fixed
+
+- build.yaml
+  - Add `Generate_Wiki_Content`, `Generate_Wiki_Sidebar`, `Clean_Markdown_Metadata`
+docs tasks. Fixes [#32](https://github.com/dsccommunity/DscResource.Base/issues/32).
 
 ## [1.3.0] - 2025-03-15