+++ CHANGELOG.md
@@ -24,13 +24,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     `Get-DscProperty` when the property `FeatureOptionalEnums` is set to
     `$true`.
   - Remove calls to `Assert()` and `Normalize()` in `Test()` and `Set()` fixes [#35](https://github.com/dsccommunity/DscResource.Base/issues/35).
-  - Add check for properties not being `$null` fixes [#30](https://github.com/dsccommunity/DscResource.Base/issues/30).
 
 ### Fixed
 
 - build.yaml
   - Add `Generate_Wiki_Content`, `Generate_Wiki_Sidebar`, `Clean_Markdown_Metadata`
 docs tasks. Fixes [#32](https://github.com/dsccommunity/DscResource.Base/issues/32).
+- `ResourceBase`
+  - Add check for properties not being `$null` fixes [#30](https://github.com/dsccommunity/DscResource.Base/issues/30).
+  - Comment typo's.
 
 ## [1.3.0] - 2025-03-15