+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2025-05-24
+
 ### Removed
 
 - Removed `Clear-ZeroedEnumPropertyValue` as it was moved and implemented